Update mod.freemember.php
This is probably a more elegant solution as it doesn't require that the user submits an extra parameter to enable query strings, instead it just persists them if they are present. I picked up on the code which Mo Variables uses to check for query strings to ensure it's robust.