Should allow optional use of 'allow' parameter
You can optionally specify an 'allow' param in both register and update tags to enforce the allowance of only specified fields in a form submission. On detection of a disallowed field (by injection for example) the user will receive a general error on submission and the attempt will be recorded in the developer log with the field name in question.

If 'allow' is used then any specified 'required' fields are also evaluated to negate double entry. :)